[ZEPPELIN-1391][Interpreters] print error while existing registedInterpreter with the same key but different settings
### What is this PR for?
print error while existing registedInterpreter with the same key but different settings.
In order to compare the property Map in the Interpreter easily,
I override the `equals` method in `InterpreterProperty` class. (and `hashCode` at the same time)

and I fix a small error in  `InterpreterProperty.toString()`, its result forgot `}` in the end.

### What type of PR is it?
Improvement

### Todos

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1391

### How should this be tested?
Existing tests.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: WeichenXu <WeichenXu123@outlook.com>

Closes #1382 from WeichenXu123/interpreter_key_error_check and squashes the following commits:

4ef032e [WeichenXu] print error while existing registedInterpreter with the same key but different settings