[ZEPPELIN-1142] Zeppelin allows two users to simultaneously edit the notebook permissions
### What is this PR for?
This is about storing the owner information on creating a note so when the same user tries to edit the permissions of the note he could do it successfully.

### What type of PR is it?
Improvement

### Todos
NA

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1142

### How should this be tested?
Check the below cases via the web application or REST API calls and the owner information should be persisted properly.
* Create note
* Clone note
* Import note

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Kavin <kavin.kumar@imaginea.com>

Closes #1393 from kavinkumarks/zeppelin-1142-simultaneous-note-permission-error and squashes the following commits:

5a45c9e [Kavin] Ensure that the authentication instance is created only when the input prinicipal is not null.
7642f63 [Kavin] Removed the duplicate instance of notebook variable and reused the existing one.
e1b8b08 [Kavin] Store owner information on creating a note and added integration test cases for the relevant scenarios.